+++ client/snapshot/delete.go
@@ -3,10 +3,8 @@ package snapshot
 import (
 	"strconv"
 
-	"github.com/spf13/cobra"
-
-	"github.com/cosmos/cosmos-sdk/client"
 	"github.com/cosmos/cosmos-sdk/server"
+	"github.com/spf13/cobra"
 )
 
 func DeleteSnapshotCmd() *cobra.Command {
@@ -15,7 +13,7 @@ func DeleteSnapshotCmd() *cobra.Command {
 		Short: "Delete a local snapshot",
 		Args:  cobra.ExactArgs(2),
 		RunE: func(cmd *cobra.Command, args []string) error {
-			viper := client.GetViperFromCmd(cmd)
+			ctx := server.GetServerContextFromCmd(cmd)
 
 			height, err := strconv.ParseUint(args[0], 10, 64)
 			if err != nil {
@@ -26,7 +24,7 @@ func DeleteSnapshotCmd() *cobra.Command {
 				return err
 			}
 
-			snapshotStore, err := server.GetSnapshotStore(viper)
+			snapshotStore, err := server.GetSnapshotStore(ctx.Viper)
 			if err != nil {
 				return err
 			}