+++ client/snapshot/dump.go
@@ -3,16 +3,13 @@ package snapshot
 import (
 	"archive/tar"
 	"compress/gzip"
-	"errors"
 	"fmt"
 	"io"
 	"os"
 	"strconv"
 
-	"github.com/spf13/cobra"
-
-	"github.com/cosmos/cosmos-sdk/client"
 	"github.com/cosmos/cosmos-sdk/server"
+	"github.com/spf13/cobra"
 )
 
 // DumpArchiveCmd returns a command to dump the snapshot as portable archive format
@@ -22,8 +19,8 @@ func DumpArchiveCmd() *cobra.Command {
 		Short: "Dump the snapshot as portable archive format",
 		Args:  cobra.ExactArgs(2),
 		RunE: func(cmd *cobra.Command, args []string) error {
-			viper := client.GetViperFromCmd(cmd)
-			snapshotStore, err := server.GetSnapshotStore(viper)
+			ctx := server.GetServerContextFromCmd(cmd)
+			snapshotStore, err := server.GetSnapshotStore(ctx.Viper)
 			if err != nil {
 				return err
 			}
@@ -49,10 +46,6 @@ func DumpArchiveCmd() *cobra.Command {
 			snapshot, err := snapshotStore.Get(height, uint32(format))
 			if err != nil {
 				return err
-			}
-
-			if snapshot == nil {
-				return errors.New("snapshot doesn't exist")
 			}
 
 			bz, err := snapshot.Marshal()
@@ -85,9 +78,27 @@ func DumpArchiveCmd() *cobra.Command {
 
 			for i := uint32(0); i < snapshot.Chunks; i++ {
 				path := snapshotStore.PathChunk(height, uint32(format), i)
-				tarName := strconv.FormatUint(uint64(i), 10)
-				if err := processChunk(tarWriter, path, tarName); err != nil {
-					return err
+				file, err := os.Open(path)
+				if err != nil {
+					return fmt.Errorf("failed to open chunk file %s: %w", path, err)
+				}
+				defer file.Close()
+
+				st, err := file.Stat()
+				if err != nil {
+					return fmt.Errorf("failed to stat chunk file %s: %w", path, err)
+				}
+
+				if err := tarWriter.WriteHeader(&tar.Header{
+					Name: strconv.FormatUint(uint64(i), 10),
+					Mode: 0o644,
+					Size: st.Size(),
+				}); err != nil {
+					return fmt.Errorf("failed to write chunk header to tar: %w", err)
+				}
+
+				if _, err := io.Copy(tarWriter, file); err != nil {
+					return fmt.Errorf("failed to write chunk to tar: %w", err)
 				}
 			}
 
@@ -106,31 +117,4 @@ func DumpArchiveCmd() *cobra.Command {
 	cmd.Flags().StringP("output", "o", "", "output file")
 
 	return cmd
-}
-
-func processChunk(tarWriter *tar.Writer, path, tarName string) error {
-	file, err := os.Open(path)
-	if err != nil {
-		return fmt.Errorf("failed to open chunk file %s: %w", path, err)
-	}
-	defer file.Close()
-
-	st, err := file.Stat()
-	if err != nil {
-		return fmt.Errorf("failed to stat chunk file %s: %w", path, err)
-	}
-
-	if err := tarWriter.WriteHeader(&tar.Header{
-		Name: tarName,
-		Mode: 0o644,
-		Size: st.Size(),
-	}); err != nil {
-		return fmt.Errorf("failed to write chunk header to tar: %w", err)
-	}
-
-	if _, err := io.Copy(tarWriter, file); err != nil {
-		return fmt.Errorf("failed to write chunk to tar: %w", err)
-	}
-
-	return nil
 }