+++ client/snapshot/restore.go
@@ -4,27 +4,22 @@ import (
 	"path/filepath"
 	"strconv"
 
-	dbm "github.com/cosmos/cosmos-db"
 	"github.com/spf13/cobra"
 
-	corestore "cosmossdk.io/core/store"
-	"cosmossdk.io/log"
-
-	"github.com/cosmos/cosmos-sdk/client"
 	"github.com/cosmos/cosmos-sdk/server"
 	servertypes "github.com/cosmos/cosmos-sdk/server/types"
+	dbm "github.com/tendermint/tm-db"
 )
 
 // RestoreSnapshotCmd returns a command to restore a snapshot
-func RestoreSnapshotCmd[T servertypes.Application](appCreator servertypes.AppCreator[T]) *cobra.Command {
+func RestoreSnapshotCmd(appCreator servertypes.AppCreator) *cobra.Command {
 	cmd := &cobra.Command{
 		Use:   "restore <height> <format>",
 		Short: "Restore app state from local snapshot",
 		Long:  "Restore app state from local snapshot",
 		Args:  cobra.ExactArgs(2),
 		RunE: func(cmd *cobra.Command, args []string) error {
-			cfg := client.GetConfigFromCmd(cmd)
-			viper := client.GetViperFromCmd(cmd)
+			ctx := server.GetServerContextFromCmd(cmd)
 
 			height, err := strconv.ParseUint(args[0], 10, 64)
 			if err != nil {
@@ -35,13 +30,12 @@ func RestoreSnapshotCmd[T servertypes.Application](appCreator servertypes.AppCre
 				return err
 			}
 
-			home := cfg.RootDir
-			db, err := openDB(home, server.GetAppDBBackend(viper))
+			home := ctx.Config.RootDir
+			db, err := openDB(home, server.GetAppDBBackend(ctx.Viper))
 			if err != nil {
 				return err
 			}
-			logger := log.NewLogger(cmd.OutOrStdout())
-			app := appCreator(logger, db, nil, viper)
+			app := appCreator(ctx.Logger, db, nil, ctx.Viper)
 
 			sm := app.SnapshotManager()
 			return sm.RestoreLocalSnapshot(height, uint32(format))
@@ -50,7 +44,7 @@ func RestoreSnapshotCmd[T servertypes.Application](appCreator servertypes.AppCre
 	return cmd
 }
 
-func openDB(rootDir string, backendType dbm.BackendType) (corestore.KVStoreWithBatch, error) {
+func openDB(rootDir string, backendType dbm.BackendType) (dbm.DB, error) {
 	dataDir := filepath.Join(rootDir, "data")
 	return dbm.NewDB("application", backendType, dataDir)
 }