+++ server/types/app.go
@@ -15,6 +15,7 @@ import (
 	"github.com/cosmos/cosmos-sdk/client"
 	"github.com/cosmos/cosmos-sdk/server/api"
 	"github.com/cosmos/cosmos-sdk/server/config"
+	"github.com/cosmos/cosmos-sdk/snapshots"
 	sdk "github.com/cosmos/cosmos-sdk/types"
 )
 
@@ -53,6 +54,9 @@ type (
 
 		// Return the multistore instance
 		CommitMultiStore() sdk.CommitMultiStore
+
+		// Return the snapshot manager
+		SnapshotManager() *snapshots.Manager
 	}
 
 	// ApplicationQueryService defines an extension of the Application interface